Override hostname on kubelet
kubelet gets the hostname from the system which on coreos does not
include the domain name (it's the short hostname and not the FQDN). This
causes the `kubernetes.io/hostname` label to mismatch the node name and
is problematic when using Service Topology and potentially other
features that use this label.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>