etcd: stagger reboot windows (#149)
Locksmithd will trigger a reboot as soon as the update-engine signals
that an update has been downloaded. Without a process to orchestrate
these reboots, there's a chance that the reboots could overlap and cause
an availability issue for the etcd cluster.

This doesn't seem to be an issue in AWS/GCP, probably because the VMs
reboot quickly, making an overlap uncommon. In Merit, however, nodes can
take up to 5 or 10 minutes to reboot, making overlaps a likelihood.

Rather than use the etcd-lock strategy, which doesn't seem to be
compatibile with our version of etcd, we can avoid clashes by assigning
each node its own daily reboot window.